Refactor generate to handle out-of-order skin tones
This is needed for Unicode v17.0. Previously, the code required all skin
tone variants of an emoji to immediately follow the base emoji in the
file. Now, we now group emojis by parsing the base name from the name of
the emoji, which all follow a consistent naming scheme. (We could have
also done it based on code points but that ended up being way more
complicated.)